refactor and solve eighteen
