+++ python/set1/seven.py
@@ -1,17 +1,6 @@
-from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
-from cryptography.hazmat.backends import default_backend
 import binascii
 
-
-def decryptEcb(ct, key):
-    defaultBackend = default_backend()
-
-    cipher = Cipher(algorithms.AES(key), modes.ECB(), backend=defaultBackend)
-    decryptor = cipher.decryptor()
-
-    pt = decryptor.update(ct) + decryptor.finalize()
-    return pt
-
+from python.common.aes import decryptEcb
 
 if __name__ == '__main__':
     key = b'YELLOW SUBMARINE'
